+++ packages/triple-web/src/env/types.ts
@@ -39,4 +39,8 @@ export interface EnvValue {
    * 미디어 소스 이름을 의미하며, 모든 측정 링크에서 반드시 포함되어야 할 유일하고 중요한 파라미터입니다.
    */
   afOnelinkPid: string
+  /**
+   * asset 파일의 웹 URL
+   */
+  webAssetsUrl?: string
 }